apps/ui/requirements.txt
pygame==2.5.2
numpy==1.26.4
paho-mqtt<2.0
orjson==3.11.3
websockets>=12.0
